refactor(payments): lift order-summary into shared lib/payments namespace
Pure relocation — no behavior change. The function is gateway-agnostic
(takes a rental_code, returns receipt data). lib/shopeepay/order-summary
becomes a re-export shim so existing imports keep working unchanged.
Shim is safe to delete after ShopeePay decommission.
